refactor: isolate inference diagnostic config
